adding utils for hosted service automation
